Update ERC20_Specifications.txt
Updated perpetual mint amounts.
Reason: I was reviewing the DAO rewards system and figured out a way to solve an old problem, that being the DAO needs to have "dynamic rewards" that reflect the cost of each Peng, and also capitalize on growing liquidity in the exchange. 
For this reason the token needs to mint more to the DAO. Reducing the exchange mint is already discussed, most of those tokens will never be utilized.